Fix static-analysis failure due to caniuse-lite dependency
As seen in https://gitlab.com/gitlab-org/gitlab/-/jobs/394057944,
static-analysis is failing with:

```
Browserslist: caniuse-lite is outdated. Please run next command
`yarn upgrade`
```

The problem may be occurring because browserslist is detecting that
caniuse-list is approximately 6 months old and outputs a warning using
console.warn.

To fix this, we change static-analysis to filter this warning message
when deciding to exit with error code 2. Older GitLab backports
shouldn't fail because we need to upgrade this dependency.

Closes https://gitlab.com/gitlab-org/gitlab/issues/194788