Clear context when scheduling mirroring jobs
The `UpdateAllMirrorsWorker` could get scheduled from a request (from
a webhook) for a certain project.

This means that it would get the pointing to the project the request
was performed for. But since the `UpdateAllMirrorsWorker` could
schedule updates for other projects, we need to clear that metadata so
it doesn't propagate to those jobs that are being scheduled.