Add current_user check in events api
The API returns a 500 error when scope is supplied and current_user
is not present. This is because a user's projects are used in the
query that returns all events.

This change adds a check to confirm that current_user is provided
before trying to run the query when scope is passed