Ensure When userWithId Is Picked User IDs are Sent
Before, userWithId would send a strategy with an empty set of parameters
as well as a strategy with the correct set of parameters.

This ensures that the primary strategy is created correctly, as well as
ensuring supplementary user IDs are added if needed.

It also removes some old code from when the user IDs were an Array.