Add documentation around plural control flow
Previously we have had cases of the `n__` method being used to control
which different strings should show up in the singular and plural cases
for a string. This doesn't work in languages such as Chinese Simplified,
as this language only has one target plural form - which means the
translator needs to determine which string should show up.

This is particularly an issue when different variables are used in each
version of the string, as variables present in the source string for
one case may not show up in the target string for the chosen plural form
from the translators - which will cause linting to fail.