Add Wiki support for AsciiDoc include directive
GitLab added support for the AsciiDoc `include` directive in 3f5d7c7.
The content of a file could be included if the file was part of the same
repository.

The Wiki `show` view was missing the expected variables needed for the
included file to be found in the repository. An AsciiDoc `include`
directive would instead cause a 500 error to be raised.

https://gitlab.com/gitlab-org/gitlab/issues/195625