Create LfsObjectsProject record for forks as well
Previous behavior was `LfsObjectsProject` are created on the
source project.

To fix the issue whenever the source project gets deleted or
fork was selectively synced, we also need to create the records
for forks.

This way, they'll also have access to the `LfsObject` and no
longer dependent on the source project.