Remove unnecessary assertions from end-to-end test
The removed assertions sometimes cause this test to be flaky, due
to the automatic discussions not being rendered in time, and so,
it's better not to have them and have this test providing reliable
feedback, than have extra assertions that cause flakiness.

In any case, we'll still have the assertions that the issue was
added/removed to/from the epic, in the epic page.

Finally, this change save us five lines of code.