Run `bin/rake pngquant:compress`
This commit is the result of running `bin/rake pngquant:compress`. All
PNG files in the doc/ directory are now optimally compressed using
pngquant.

This is the first time this command has been run on this codebase, so in
the future this command will not produce nearly as many changes.

This commit also bumps up the compression threshold of the
pngquant:compress task to avoid changing too many files.