Rename RelativeLinkFilter -> RepositoryLinkFilter
The end goal of this change is to split the RelativeLinkFilter into two
filters:

1. UploadLinkFilter (must run first)
2. RepositoryLinkFilter

This means that each filter will be easier to understand, and we'll also
be able to apply Gitaly-specific optimisations to the
RepositoryLinkFilter without having to worry about upload links. By
putting the UploadLinkFilter first, we'll also effectively short-circuit
any Gitaly calls when a document's only relative links are upload links.

This is a standalone commit because it's a rename and those can be hard
to follow when mixed with other changes.