Bails out duplicated entries on namespaces worker
There seems to be some sort on race when creating records on the
namespaces_aggregation_schedules table. Probably due to multiple workers
trying to update the statistics of the same namespace. Since the error
is harmless, we decided to catch the exception and bailout.

Related to https://gitlab.com/gitlab-org/gitlab/issues/195427