Add types to style guide
`_placeholder` refers to a temporary element that appears while
content is loading. For example, the elements that are displayed
instead of discussions while the discussions are being fetched.

`_container` refers to an element that includes other elements, but
might not include any visible content itself. E.g., an element that
has a third-party editor inside it, but which isn't the editor itself
and so doesn't include the editor's content.

`_modal` refers to a popup modal dialog, e.g., a confirmation prompt.