Remove extra assertions from end-to-end test
By asserting that the epic title is the same of the issue that
was promoted we are already testing that issue promotion works, so
better to remove this extra assertions that might cause test flaki-
ness since sometimes automatic discussions take longer to render.

Also, by removing an extra step + assertion, we make the test a bit
faster.

Finally, this way the code gets a bit shorter.