Add subscription_plan and caller_id to ApplicationContext
It adds new attributes subscription_plan and caller_id.

The subscription_plan can be "free", "bronze", "silver" or "gold",
these are exclusively fetched from the given namespace (or the
project if it's not given).

The caller_id can be the controller action name if it's a regular
Rails request, or the Grape route, when it's an API request.

These new attributes will also be logged at Sidekiq logs, along
the project, user and namespace.