Set default reconnect_attempts of Rails.cache to 1
RedisCacheStore defaults to 0 but we need this to be 1 because
Redis connections need to reconnect after a fork so that forked
processes don't use the same socket

This problem was introduced when we switched to RedisCacheStore
because Rails explicitly sets the default to 0. The default value
in the redis client is 1.