Prevent RecordsUploads from recording versions
Previously, versions of images would be recorded to the `uploads` table
in addition to the main (unresized) image also being recorded.

This change prevents the versions from being recorded in the uploads
table.

Note that this is an incremental piece of work for
https://gitlab.com/gitlab-org/gitlab/issues/22693

See https://gitlab.com/gitlab-org/gitlab/issues/22693#note_269469039
for a description of further steps.