Use first and last name for new sign up flow
With the new paid sign up flow, we will create a customer on the
customer portal. The customer portal requires a first and last name.
We already store first and last name in the database.

Since the new sign up design  will in the future also be used for
the FOSS version of GitLab, we need to accept first and last name
there as well.