(ISSUE35690) unquarantine test
We are unquarantining this test for two reasons.
 1) Because of
https://gitlab.com/gitlab-org/gitlab/merge_requests/22525 has
not been merged yet, we cannot run this test in quarantine and we need
to run the test to determine if it is still flaky, and if it is we can
then examine the artifacts.
 2) Because we want to know if Sanad's changes have fixed this test
https://gitlab.com/gitlab-org/gitlab/
commit/d58847b735c398f7675e3f8d7d6e8141c0a932e0