Fix broken shallowMount with GlModal
**How?**
Making this spec a full mount instead.

**Why?**
The test was failing with an error in our
GlModal that `$refs.modal.show` was not
a function. This is because shallowMount
will stub the internal b-modal. So that we're
not messing with internals too much, let's just
do a full mount here :)