Generate smaller image sizes for Designs
Adds a new DesignV432x230Uploader to generate and serve smaller
versions of design files.

The original (full-sized) design files are stored in Git LFS, and so
have a different uploader, `LfsObjectUploader`.

The decision to make a separate uploader to handle just the smaller
image versions came out of a discussion in
https://gitlab.com/gitlab-org/gitlab/merge_requests/20511. Initially,
the direction taken was to make LfsObjectUploader conditionally generate
smaller images for LfsObject files. The feedback was that this was
adding logic that was too specific to design management into an uploader
that is very general, and that we should use a separate uploader to
handle just the smaller image versions.

A service now runs after a new DesignManagement::Version has been
created, which downloads the original design files from their Git LFS
storage and generates smaller versions, storing them in the regular
uploads storage.

This change is behind the :design_management_resize_images feature flag.

https://gitlab.com/gitlab-org/gitlab/issues/12577
https://gitlab.com/gitlab-org/gitlab/issues/13815