Display message if MR Security report is outdated
When the last pipeline for the target branch hadn't detected some
security issues, maybe due to jobs failed or artifacts expired, the
MR security report should be considered outdated and the pipeline needs
to be re-run. It also handles the edge case when target branch has
received more commits and the current branch needs to be updated.