Add Rake task to generate all_queues.yml
The validity of this file is already checked by
spec/workers/every_sidekiq_worker_spec.rb. This adds a second, simple
check that the file loads as YAML matching the Rake task's generated
list (including sorting).

The files generated by the task are in broadly the same format as the
current manually-written files:

1. Namespaced queues first, then un-namespaced queues.
2. Both in alphabetical order.

They don't have the manually-inserted line breaks. This doesn't delete
the RSpec test because that also adds an important check for the
addition of two queues, mailers and default, that shouldn't be lost in
any refactor here.