Add EE GFM pipeline spec
This adds a spec for the EE GFM pipeline.

Test examples are added here to check that:
 * performance is acceptable
 * overlapping patterns do not cause errors

As an incidental improvement, it was found that there was a bug in the
issue reference filter, which is fixed by renaming a method to fix
the incorrect naming of relation_for_paths.