Gracefully handle marking a project deletion multiple times
If a user clicks on "Delete repository" on a project that has already
been marked for deletion, GitLab would previously throw a 500 Error due
to `NoMethodError: undefined method '[]' for nil:NilClass`. To fix this,
return a successful mark so that the user can see when the project is
slated for deletion.

Closes https://gitlab.com/gitlab-org/gitlab/issues/121730