Add epic in filtered search
This MR adds a new option in the filtered search: Epic. With this
change, when the hint dropdown opens, epic will be an option. Upon
selecting, all the epics in the parent group will be displayed. When an
epic is selected, the only issues contained in the epic are displayed.
This also supports the new NOT operator.