Fix subscription generation incorrect trial
When namespace does not have subscription, it is generated on-the-fly.
The trial column is supposed to be based on deprecated trial_ends_on
column, but not trial_ends_on is delegated to gitlab_subscription,
which is nil (and a circular dependency).

The fix is to fetch trial_ends_on from namespaces table directly.