Limit page number on explore/projects
Limit the maximum page number to 500 on project exploration.
Prior to this change it was possible to iterate infinitely, and
high page numbers are prone to slow speeds due to the project list
being a slow query amplified by OFFSET pagination slowness.

There's a new Prometheus metric available at
`gitlab_page_out_of_bounds` for monitoring the traffic which hits the
page limit.