Update name max length
To support the same amount of consecutive characters as before in a name
we limit `first_name` and `last_name` to 127 characters each and limit
`name` to 255 (127 * 2 + 1 whitespace character).

Although `first_name` and `last_name` would both support up to 255
characters in the database, we have a limit of `255` characters on the
name of a `Namespace` and `username`. The former gets generated out of
the name and the latter gets suggested to a user when creating a trial