Fix broken FOSS build in import_test_coverage_spec.rb
97f102f46b72a91dfbe879a2b20f913ee014e0d0 introduced a fixture that is
only present in EE. When the FOSS build attempt to run this test, it
failed because the file did not exist.

To fix this, we only add this fixture under an EE build.

We also need to mute a FOSS-only failure with `project.issues.events`
because that relation is only tested in the EE-specific fixture.

Closes https://gitlab.com/gitlab-org/gitlab/issues/197298