Use the slug as the project name if empty
Added on change events to the project path so only once the slug has
been created will we generate the project name (since we don't want
to change the project name if the slug changes again).

On the git URL importer the slug/project names are more dynamic
(based on the URL given).

Updated some old-style jQuery event references and updated the
humanize method to accept a separator.

Fixed gitlab importer so it actually copies between fields.
Updated to convert the slug to title case and added test cases.

Fixed slug edge-cases with emoji's in the project name.

Updated documentation.

Added changelog