Rename identifier_for_subject to identifier_for_repositorable
In https://gitlab.com/gitlab-org/gitlab/issues/39176 we're
allowing resources other than projects to have repositories
attached.

Therefore, we need to refactor those places where we use
project because it can also be a snippet. Besides, we need to
abstract this variable to something like repositorable.

In this commit we rename the method
GlRepository::RepoType#identifier_for_subject to
identifier_for_repositorable to align with this.