Fix two feature specs that fail with CHROME_HEADLESS=0
In the Web IDE, compact mode depends on the size of window.
If it is shorter than MAX_WINDOW_HEIGHT_COMPACT, (as it is
with CHROME_HEADLESS=0), the initial commit button will exist.

Otherwise, if it is taller (as it is by default with chrome headless)
then the button will not exist.