Move Phabricator import workers to correct location
These should be in app/workers. If they aren't, they won't be picked up
by every_sidekiq_worker_spec.rb, and their queues won't be guaranteed to
be in all_queues.yml, which in turn prevents sidekiq-cluster from seeing
them.

Making the BaseWorker not include ApplicationWorker means that we don't
need to act as if it has its own queue. (See HashedStorage::BaseWorker,
which does the same thing.)