Make merge trains in tests independent
If multiple tests use the same project they'll use the same merge
train. That can cause tests to fail if one tries to use the merge
train while it's in an unexpected state, e.g., trying to start a new
train when one is already running.

This creates a separate project for each test that uses a merge
train.

Uses the Commit API to add a file because it's faster than
pushing.

Adds a runners_token attribute to the Group resource, and lets
resources be reloaded so updated data can be fetched. It's not
clear why the runners token isn't available immediately upon
creating a group, but reloading the API data fixes it.