Document API desc block when released behind a FF
We ask to document on which version a new endpoint is released, but when
using a feature flag is easy to get it wrong.

Here we specify what to write in such case, so that it will be easier to
set the correct GitLab version once we remove the feature flag.