Fix: incorrect security status counts
In the project security status widget on the group security dashboard,
we were showing the count of vulnerabilities found in all default
branch pipelines, but we *should* have been showing only the count
for the latest pipeline.

This commit updates our query to only fetch vulnerabilities from the
latest pipeline.

https://gitlab.com/gitlab-org/gitlab/issues/119020