Import/export snippet repositories
When the snippet has a repository attached,
the export file will create a bundle per each
one.

When we restore the export file, if the snippet has
a repository, we will import it from the bundle.
If not, we will create a repository from the
snippet file_name and content columns.