Stub DNS in the spec suite
This implements stubbing throughout the spec suite for DNS queries,
of which there are many. It includes a few tweaks to continue to allow
resolution for local addresses. This solves an issue where non-standard
DNS setups prevent most of the spec suite from working locally, and
should also provide speed improvements to the CI pipeline on average
by removing a source of external requests.