Trim extra period when merge error displayed
When a merge error happens, previously the UI displayed two periods:

```
Merge failed: Failed to squash. Should be done manually.. Please try
again.
```

This happened because the backend sent the error message, "Failed to
squash. Should be done manually.", while the frontend added an extra
period.

To avoid this issue, we now trim the period and add it in the sentence.