Improve error handling of squash and rebase
Previously all Git-related exceptions in squash and rebase would quietly
go to `git_json.log`, but these messages should really be tracked in
Sentry and/or `exceptions_json.log`.

In addition, if the squash in progress check failed, the merge would get
stuck. This commit now handles that error gracefully and logs it.

In the future, we can likely remove `git_json.log` altogether.