Add Vale as an English language linter
Vale (https://errata-ai.github.io/vale/) is a grammar, style, and word
usage linter for the English language. This merge request adds an
INI file for Vale (.vale.ini) to the root directory of the repository,
and creates a hidden folder (doc/.linting) for the storage of linting
tests.

Since the TW team uses multiple linting tools, and Vale also supports
multiple collections of rules (called 'styles'), the path to the tests
folder is doc/.linting/vale/styles/gitlab to support any future
expansions that can't be predicted at this time.

The folder currently contains only a .gitkeep file to meet the goal
of a MVC. Tests should be developed, vetted, and added in separate
merge requests, but those MRs all depend on the addition of the
INI file and the directory structure.