Add specs for Gitlab::SidekiqConfig::CliMethods
By using `fast_spec_helper`, we also do a better job of testing that
these methods don't expect un-required files to be
available. `.worker_queues` is tested here and in the specs for
Gitlab::SidekiqConfig, as in the latter it's using the actual YAML
files, rather than stubbed ones as added here.