Create and write to deployment_clusters table
Creates a table "deployment_clusters" to expand the cluster-related
deployment details from just cluster_id to include kubernetes namespaces
as well, and adds double writes to this table. In the next
release, we can then copy the remaining data over, start reading
exclusively from deployment_clusters, and add an ignore for
deployments.cluster_id

Storing the namespace used is important for accurate deploy boards,
because Environment#deployment_namespace is the namespace for the *next*
deployment, and this can change as cluster configuration changes.