Refactor project destroy services
In preparation for
https://gitlab.com/gitlab-org/gitlab/merge_requests/22672,
we need to refactor Projects::DestroyService. At the
moment, this service has hardcoded the repositories it
has to delete.

Nevertheless, we have now more requirements that
we had before. We need to remove Snippet repositories
and also DesignManagement ones.

In this commit, we refactor the existing
Projects::DestroyService creating two services just for
dealing with repositories and extracting this
functionality from this service.

We also split some of the existing functionality. The
current destroy service delete things but also is able
to rollback stuff. We can move this functionality into
a different service.