Prevent DOS via excessive ASCIIDOC includes
This sets a maximum limit on the number of include directives that any
one file may use (transitively). This is set relatively low at the
moment to 32 since each one requires a gitlay call to retrieve a blob
from the repo, and experimentation suggests that this keeps processing
to under 10sec.

This approach prevents both cyclic imports as well as having a single
file with an extremely lang list of includes.