Remove deprecated caching class
With the launch of the new approach to caching highlighting diffs in
redis, we can drop this class as it is no longer needed. We used it in
the new highlighting class to accelerate the expiration of deprecated
cache entries, but as of 12.7 the new caching approach is enabled by
default, allowing us to completely remove this.