Create a Vale rule for the word Gitlab
Using "GitLab's" as a shortened form of "GitLab is" or "GitLab was"
is not acceptable in our style guide, but "GitLab's" as a
possessive form is not acceptable. Since no automated test can
distinguish between the two usages, this commit creates a test
that provides a low-priority result encouraging the user to
check the usage.