Add logic to create a cluster management project
This logic isn't currently used by any process, but
eventually a management project will be created when
a user installs an application to the cluster (if
there is no management project already specified).
This will usually be when installing the first
application to the cluster, but could also be on
subsequent installs if the project is somehow unset
or deleted.

The project is currently created empty, the next
step will be to populate it with a sample cluster
management repository.