Part 2 of project milestone resource refactoring
This commit moves attributes of a project milestone resource to the
main class so that the EE one can inherit it, and focus only on EE
specific stuff.

And one of the attributes is being removed since it wasn't in use.

Also, it prefers to be explicit about how fabrication happens (via
API) than being implicit (fabricate!)

Finally, this change saves us six lines of code.