Use FLOAT_TYPE for possibly large fields.
This does update the schema, but does not require large updates in the frontend, as opposed to Types::BigInt which would require string -> number casting. 
GraphQL-induced type casting should prevent type issues according to https://spec.graphql.org/June2018/#sec-Float

The values allowed in this type should suffice for the foreseeable future.

Also  updated the changelog to document this limitation.