Run feature specs through Workhorse
In development and production, Workhorse is an essential part of how
GitLab operates. However, it is currently excluded from tests. This
makes it hard to approach features that rely on workhorse - like file
uploads - in a test-driven manner.

Introducing workhorse into the test suite for feature specs seems to
have minimal speed impact, and should help significantly with this.