Specify syntax highlighting for shell output
For the sake of consistency and simplicity, we're specifying to use
'shell' instead of 'bash' or 'sh' for most syntax highlighting most
shell output. The exception: code blocks that contain `sudo`,
because Mike Jang discovered it causes oddities in the syntax
highlighting output. For code blocks containing `sudo`, we're now
specifying to use `console` instead.`