Add a Rails backtrace silencer for EE load balancing code
Prior to this change, the SQL backtrace in development.log and test.log
showed `ee/lib/gitlab/database/load_balancing/connection_proxy.rb:63` as
the caller of many queries, which is not helpful in debugging. Filter
these lines out so we don't see them there.

In production, we may want to see these lines in case there is a bug in
the load balancer.