Revert broken background  migration
This revert PatchPrometheusServicesForSharedClusterApplications
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/22974
because it has the potential of blowing up Sidekiq queues.
Background jobs are intended to process many rows for
a large batch of records, but this migration tends to process
a single row within a second.

Mikolaj W: With exception for adding temporary index
as plain revert of this migration can potentially
leave some databases in inconsistent state.

Identified in
https://gitlab.com/gitlab-com/gl-infra/infrastructure/issues/9064