Drop problematic background jobs from the Sidekiq queue
This uses the Sidekiq API to drop
`PatchPrometheusServicesForSharedClusterApplications` because this job
can blow up Sidekiq queues by retrying excessively.
Background jobs are intended to process many rows for
a large batch of records, but this migration tends to process
a single row within a second. As a result, the exclusive lease causes
this job to rescheduled over and over again.

Identified in
https://gitlab.com/gitlab-com/gl-infra/infrastructure/issues/9064